arrumando quebra de layout de falta de produtos

+++ index.php
@@ -17,20 +17,22 @@
 <div class="container">
   <div class="row">
         <?php
-            foreach($products_arr['data'] as $data){
-                echo '<div class="col-4">';
-                    echo '<div class="card">';
-                        echo '<div class="card-body">';
-                            echo '<h3 class="card-title">' . $data['product_name'] . '</h3>';
-                            echo '<h6 class="card-subtitle mb-2 text-muted">Cores: </h6>';
-                                foreach($data['colours'] as $colours){
-                                    echo '<p class="card-text">'.$colours.'</p>';
-                                }
-                            echo '<a href="api/product/show_by_id.php?id=' . $data["id_product"] . '" class="card-link">Alterar</a>';
-                            echo '<a href="api/product/delete.php?id=' . $data["id_product"] . '" class="card-link">Deletar</a>';
+            if(!empty($products_arr['data'])){
+                foreach($products_arr['data'] as $data){
+                    echo '<div class="col-4">';
+                        echo '<div class="card">';
+                            echo '<div class="card-body">';
+                                echo '<h3 class="card-title">' . $data['product_name'] . '</h3>';
+                                echo '<h6 class="card-subtitle mb-2 text-muted">Cores: </h6>';
+                                    foreach($data['colours'] as $colours){
+                                        echo '<p class="card-text">'.$colours.'</p>';
+                                    }
+                                echo '<a href="api/product/show_by_id.php?id=' . $data["id_product"] . '" class="card-link">Alterar</a>';
+                                echo '<a href="api/product/delete.php?id=' . $data["id_product"] . '" class="card-link">Deletar</a>';
+                            echo '</div>';
                         echo '</div>';
                     echo '</div>';
-                echo '</div>';
+                }
             }
         ?>
     </div>
